# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ReliefMesh API
  version: 0.1.0
  summary: Offline-first, privacy-preserving coordination for local disaster-response exercises and community aid.
  description: |
    REST API of a self-hosted ReliefMesh instance.

    **ReliefMesh is not an emergency service.** For immediate danger to life, health, fire, crime or other
    emergencies, contact the responsible local emergency services. Nothing in this API notifies authorities,
    dispatches responders or performs medical triage.

    ## Authentication
    Sessions use an HTTP-only, `SameSite=Strict` cookie (`__Host-rm_session` over HTTPS, `rm_session` in local
    development). Every state-changing request from an authenticated session must send the CSRF token returned by
    `POST /auth/login` or `GET /auth/session` in the `X-CSRF-Token` header. Browsers must also send a matching
    `Origin`.

    ## Authorization
    All authorization is enforced on the server. Records the caller may not see are reported as `404` so their
    existence is not revealed. Every view contains a `permissions` object computed on the server.

    ## Privacy
    Protected values (contact details and exact locations) are never part of regular responses. They are only
    returned by the `.../protected` endpoints, which record an audit event that is visible to the data subject.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Same origin as the web app
...
security:
  - session: []
paths:
...
  /api/v1/requests/{id}/notes:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [notes]
      summary: List notes visible to the caller
      responses:
        "200": { $ref: "#/components/responses/NoteList" }
    post:
      tags: [notes]
      summary: Add an append-only note
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody: { $ref: "#/components/requestBodies/Note" }
      responses:
        "201": { $ref: "#/components/responses/Note" }
        "200": { $ref: "#/components/responses/Note" }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/offers/{id}/notes:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [notes]
      summary: List offer notes
      responses:
        "200": { $ref: "#/components/responses/NoteList" }
    post:
      tags: [notes]
      summary: Add an offer note
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody: { $ref: "#/components/requestBodies/Note" }
      responses:
        "201": { $ref: "#/components/responses/Note" }
        "200": { $ref: "#/components/responses/Note" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: __Host-rm_session
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    CSRF: { name: X-CSRF-Token, in: header, required: true, schema: { type: string } }
  requestBodies:
...
    Note:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [body]
            properties:
              client_id: { type: string, format: uuid }
              body: { type: string, minLength: 1, maxLength: 4000 }
              visibility: { type: string, enum: [internal, responders, shared] }
              is_sensitive: { type: boolean }
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    Note:
      description: Note
      content: { application/json: { schema: { $ref: "#/components/schemas/Note" } } }
    NoteList:
      description: Notes
      content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/Note" } } } } } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [validation_failed, forbidden, not_found, invalid_transition, over_allocation, version_conflict, csrf_invalid] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            details: { type: object }
            request_id: { type: string }
...
    Note:
      type: object
      properties:
        id: { type: string, format: uuid }
        visibility: { type: string, enum: [internal, responders, shared] }
        is_sensitive: { type: boolean }
        body: { type: string }
        created_at: { type: string, format: date-time }
        redacted: { type: boolean }
        author: { type: object, properties: { label: { type: string }, display_name: { type: string }, is_you: { type: boolean } } }
...